==> 01_EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data = pd.read_csv('data.csv')

len(data)

data.head(5)

data.source.value_counts()

data.source.unique()

data.source.value_counts()/len(data)

from sklearn.model_selection import train_test_split

data['text_length'].max()

data[data['text_length'] == data['text_length'].max()]

plt.figure(figsize=(10, 6))
plt.hist(data['text_length'], bins=30, color='blue', alpha=0.7)
plt.title('Distribution of Text Length')
plt.xlabel('Text Length')
plt.ylabel('Frequency')
plt.grid(True)
plt.show()


plt.figure(figsize=(10, 6))
plt.boxplot(data['text_length'], vert=False)
plt.title('Box Plot of Text Length')
plt.xlabel('Text Length')
plt.grid(True)
plt.show()


data['text_length'].describe()

percentile_95 = data['text_length'].quantile(0.95)
print("The 95th percentile of text length is:", percentile_95)

data_fil = data[data['text_length'] <= 2000]

data_fil

value_counts_series = data_fil['source'].value_counts()

# Convert to DataFrame
value_counts_df = value_counts_series.reset_index()
value_counts_df.columns = ['Unique_Value', 'Counts']  # Rename columns for clarity

print(value_counts_df)

value_counts_df[value_counts_df['Counts'] >= 1000].nunique

filter_source_list=value_counts_df[value_counts_df['Counts'] >= 1000]['Unique_Value'].unique().tolist()

data_fin=data_fil[data_fil['source'].isin(filter_source_list)]

# Split the data into training and test sets
train_df, test_df = train_test_split(data_fin, test_size=0.2, random_state=42, stratify=data_fin['source'])

# Check the split
print(f"Training set size: {len(train_df)}")
print(f"Test set size: {len(test_df)}")


# Split the data into training and test sets
train_df, val_df = train_test_split(train_df, test_size=0.2, random_state=42, stratify=train_df['source'])

# Check the split
print(f"Training set size: {len(train_df)}")
print(f"Val set size: {len(val_df)}")


train_df.to_csv('train_fil.csv')
val_df.to_csv('val_fil.csv')
test_df.to_csv('test_fil.csv')
